+++ test/Sentry.Tests/SentryHttpMessageHandlerTests.cs
@@ -36,6 +36,7 @@ public class SentryHttpMessageHandlerTests
     {
         // Arrange
         var hub = Substitute.For<IHub>();
+        var failedRequestHandler = Substitute.For<ISentryFailedRequestHandler>();
         var options = new SentryOptions
         {
             TracePropagationTargets = new List<TracePropagationTarget>
@@ -48,7 +49,7 @@ public class SentryHttpMessageHandlerTests
             SentryTraceHeader.Parse("75302ac48a024bde9a3b3734a82e36c8-1000000000000000-0"));
 
         using var innerHandler = new RecordingHttpMessageHandler(new FakeHttpMessageHandler());
-        using var sentryHandler = new SentryHttpMessageHandler(hub, options, innerHandler);
+        using var sentryHandler = new SentryHttpMessageHandler(hub, options, innerHandler, failedRequestHandler);
         using var client = new HttpClient(sentryHandler);
 
         // Act
@@ -67,6 +68,7 @@ public class SentryHttpMessageHandlerTests
     {
         // Arrange
         var hub = Substitute.For<IHub>();
+        var failedRequestHandler = Substitute.For<ISentryFailedRequestHandler>();
         var options = new SentryOptions
         {
             TracePropagationTargets = new List<TracePropagationTarget>
@@ -79,7 +81,7 @@ public class SentryHttpMessageHandlerTests
             SentryTraceHeader.Parse("75302ac48a024bde9a3b3734a82e36c8-1000000000000000-0"));
 
         using var innerHandler = new RecordingHttpMessageHandler(new FakeHttpMessageHandler());
-        using var sentryHandler = new SentryHttpMessageHandler(hub, options, innerHandler);
+        using var sentryHandler = new SentryHttpMessageHandler(hub, options, innerHandler, failedRequestHandler);
         using var client = new HttpClient(sentryHandler);
 
         // Act
@@ -215,6 +217,26 @@ public class SentryHttpMessageHandlerTests
         Assert.Equal(expectedBreadcrumbData[statusKey], breadcrumbGenerated.Data[statusKey]);
     }
 
+    [Fact]
+    public async Task SendAsync_Executed_FailedRequestsCaptured()
+    {
+        // Arrange
+        var hub = Substitute.For<IHub>();
+        var failedRequestHandler = Substitute.For<ISentryFailedRequestHandler>();
+        var options = new SentryOptions();
+        var url = "https://localhost/";
+
+        using var innerHandler = new FakeHttpMessageHandler();
+        using var sentryHandler = new SentryHttpMessageHandler(hub, options, innerHandler, failedRequestHandler);
+        using var client = new HttpClient(sentryHandler);
+
+        // Act
+        await client.GetAsync(url);
+
+        // Assert
+        failedRequestHandler.Received(1).HandleResponse(Arg.Any<HttpResponseMessage>());
+    }
+
 #if NET5_0_OR_GREATER
     [Fact]
     public void Send_SentryTraceHeaderNotSet_SetsHeader_ByDefault()
@@ -245,6 +267,7 @@ public class SentryHttpMessageHandlerTests
     {
         // Arrange
         var hub = Substitute.For<IHub>();
+        var failedRequestHandler = Substitute.For<ISentryFailedRequestHandler>();
         var options = new SentryOptions
         {
             TracePropagationTargets = new List<TracePropagationTarget>
@@ -257,7 +280,7 @@ public class SentryHttpMessageHandlerTests
             SentryTraceHeader.Parse("75302ac48a024bde9a3b3734a82e36c8-1000000000000000-0"));
 
         using var innerHandler = new RecordingHttpMessageHandler(new FakeHttpMessageHandler());
-        using var sentryHandler = new SentryHttpMessageHandler(hub, options, innerHandler);
+        using var sentryHandler = new SentryHttpMessageHandler(hub, options, innerHandler, failedRequestHandler);
         using var client = new HttpClient(sentryHandler);
 
         // Act
@@ -276,6 +299,7 @@ public class SentryHttpMessageHandlerTests
     {
         // Arrange
         var hub = Substitute.For<IHub>();
+        var failedRequestHandler = Substitute.For<ISentryFailedRequestHandler>();
         var options = new SentryOptions
         {
             TracePropagationTargets = new List<TracePropagationTarget>
@@ -288,7 +312,7 @@ public class SentryHttpMessageHandlerTests
             SentryTraceHeader.Parse("75302ac48a024bde9a3b3734a82e36c8-1000000000000000-0"));
 
         using var innerHandler = new RecordingHttpMessageHandler(new FakeHttpMessageHandler());
-        using var sentryHandler = new SentryHttpMessageHandler(hub, options, innerHandler);
+        using var sentryHandler = new SentryHttpMessageHandler(hub, options, innerHandler, failedRequestHandler);
         using var client = new HttpClient(sentryHandler);
 
         // Act
@@ -422,6 +446,26 @@ public class SentryHttpMessageHandlerTests
 
         Assert.True(breadcrumbGenerated.Data.ContainsKey(statusKey));
         Assert.Equal(expectedBreadcrumbData[statusKey], breadcrumbGenerated.Data[statusKey]);
+    }
+
+    [Fact]
+    public void Send_Executed_FailedRequestsCaptured()
+    {
+        // Arrange
+        var hub = Substitute.For<IHub>();
+        var failedRequestHandler = Substitute.For<ISentryFailedRequestHandler>();
+        var options = new SentryOptions();
+        var url = "https://localhost/";
+
+        using var innerHandler = new FakeHttpMessageHandler();
+        using var sentryHandler = new SentryHttpMessageHandler(hub, options, innerHandler, failedRequestHandler);
+        using var client = new HttpClient(sentryHandler);
+
+        // Act
+        client.Get(url);
+
+        // Assert
+        failedRequestHandler.Received(1).HandleResponse(Arg.Any<HttpResponseMessage>());
     }
 #endif
 }